Moves ledger into appState
