+++ app/ledger.js
@@ -55,8 +55,6 @@ const uuid = require('uuid')
 
 const appActions = require('../js/actions/appActions')
 const appConfig = require('../js/constants/appConfig')
-const appConstants = require('../js/constants/appConstants')
-const windowConstants = require('../js/constants/windowConstants')
 const messages = require('../js/constants/messages')
 const settings = require('../js/constants/settings')
 const request = require('../js/lib/request')
@@ -67,7 +65,6 @@ const rulesolver = require('./extensions/brave/content/scripts/pageInformation')
 const ledgerUtil = require('./common/lib/ledgerUtil')
 const tabs = require('./browser/tabs')
 const pageDataState = require('./common/state/pageDataState')
-const {fileUrl} = require('../js/lib/appUrlUtil')
 
 // "only-when-needed" loading...
 let ledgerBalance = null
@@ -76,18 +73,15 @@ let ledgerGeoIP = null
 let ledgerPublisher = null
 
 // testing data
-const testVerifiedPublishers = [
-  'brianbondy.com',
-  'clifton.io'
-]
+
 
 // TBD: remove these post beta [MTR]
+// TODO remove, it's not used anymore
 const logPath = 'ledger-log.json'
 const publisherPath = 'ledger-publisher.json'
 const scoresPath = 'ledger-scores.json'
 
 // TBD: move these to secureState post beta [MTR]
-const statePath = 'ledger-state.json'
 const synopsisPath = 'ledger-synopsis.json'
 
 /*
@@ -104,15 +98,7 @@ const clientOptions = {
   server: process.env.LEDGER_SERVER_URL,
   createWorker: app.createWorker
 }
-
-var doneTimer
 var quitP
-
-var v2RulesetDB
-const v2RulesetPath = 'ledger-rulesV2.leveldb'
-
-var v2PublishersDB
-const v2PublishersPath = 'ledger-publishersV2.leveldb'
 
 /*
  * publisher globals
@@ -126,14 +112,6 @@ var publishers = {}
  * utility globals
  */
 
-const msecs = { year: 365 * 24 * 60 * 60 * 1000,
-  week: 7 * 24 * 60 * 60 * 1000,
-  day: 24 * 60 * 60 * 1000,
-  hour: 60 * 60 * 1000,
-  minute: 60 * 1000,
-  second: 1000
-}
-
 /*
  * notification state globals
  */
@@ -144,178 +122,11 @@ let notificationPaymentDoneMessage
 let notificationTryPaymentsMessage
 let notificationTimeout = null
 
-// TODO(bridiver) - create a better way to get setting changes
-const doAction = (state, action) => {
-  var i, publisher
-
-/* TBD: handle
-
-    { actionType: "window-set-blocked-by"
-    , frameProps:
-      { audioPlaybackActive: true
-        ...
-      }
-    , ...
-    }
- */
-  if (publisherInfo._internal.debugP) {
-    console.log('\napplication event: ' + JSON.stringify(underscore.pick(action, [ 'actionType', 'key' ]), null, 2))
-  }
-
-  switch (action.actionType) {
-    case appConstants.APP_SET_STATE:
-      init()
-      break
-
-    case appConstants.APP_BACKUP_KEYS:
-      state = backupKeys(state, action)
-      break
-
-    case appConstants.APP_RECOVER_WALLET:
-      state = recoverKeys(state, action)
-      break
-
-    case appConstants.APP_SHUTTING_DOWN:
-      quit()
-      break
-
-    case appConstants.APP_ON_CLEAR_BROWSING_DATA:
-      {
-        const defaults = state.get('clearBrowsingDataDefaults')
-        const temp = state.get('tempClearBrowsingData', Immutable.Map())
-        const clearData = defaults ? defaults.merge(temp) : temp
-        if (clearData.get('browserHistory') && !getSetting(settings.PAYMENTS_ENABLED)) {
-          reset(true)
-        }
-        break
-      }
-
-    case appConstants.APP_IDLE_STATE_CHANGED:
-      state = pageDataChanged(state)
-      visit('NOOP', underscore.now(), null)
-      break
-
-    case appConstants.APP_CHANGE_SETTING:
-      switch (action.key) {
-        case settings.PAYMENTS_ENABLED:
-          initialize(action.value)
-          break
-
-        case settings.PAYMENTS_CONTRIBUTION_AMOUNT:
-          setPaymentInfo(action.value)
-          break
-
-        case settings.PAYMENTS_MINIMUM_VISIT_TIME:
-          if (action.value <= 0) break
-
-          synopsis.options.minPublisherDuration = action.value
-          updatePublisherInfo()
-          break
-
-        case settings.PAYMENTS_MINIMUM_VISITS:
-          if (action.value <= 0) break
-
-          synopsis.options.minPublisherVisits = action.value
-          updatePublisherInfo()
-          break
-
-        case settings.PAYMENTS_ALLOW_NON_VERIFIED:
-          synopsis.options.showOnlyVerified = action.value
-          updatePublisherInfo()
-          break
-
-        default:
-          break
-      }
-      break
-
-    case appConstants.APP_CHANGE_SITE_SETTING:
-      if (!action.hostPattern) {
-        console.warn('Changing site settings should always have a hostPattern')
-        break
-      }
-      i = action.hostPattern.indexOf('://')
-      if (i === -1) break
-
-      publisher = action.hostPattern.substr(i + 3)
-      if (action.key === 'ledgerPaymentsShown') {
-        if (action.value === false) {
-          if (publisherInfo._internal.verboseP) console.log('\npurging ' + publisher)
-          delete synopsis.publishers[publisher]
-          delete publishers[publisher]
-          updatePublisherInfo()
-        }
-      } else if (action.key === 'ledgerPayments') {
-        if (!synopsis.publishers[publisher]) break
-
-        if (publisherInfo._internal.verboseP) console.log('\nupdating ' + publisher + ' stickyP=' + action.value)
-        updatePublisherInfo()
-        verifiedP(publisher)
-      } else if (action.key === 'ledgerPinPercentage') {
-        if (!synopsis.publishers[publisher]) break
-        synopsis.publishers[publisher].pinPercentage = action.value
-        updatePublisherInfo(publisher)
-      }
-      break
-
-    case appConstants.APP_REMOVE_SITE_SETTING:
-      i = action.hostPattern.indexOf('://')
-      if (i === -1) break
-
-      publisher = action.hostPattern.substr(i + 3)
-      if (action.key === 'ledgerPayments') {
-        if (!synopsis.publishers[publisher]) break
-
-        if (publisherInfo._internal.verboseP) console.log('\nupdating ' + publisher + ' stickyP=' + true)
-        updatePublisherInfo()
-      }
-      break
-
-    case appConstants.APP_NETWORK_CONNECTED:
-      setTimeout(networkConnected, 1 * msecs.second)
-      break
-
-    case appConstants.APP_NAVIGATOR_HANDLER_REGISTERED:
-      ledgerInfo.hasBitcoinHandler = (action.protocol === 'bitcoin')
-      appActions.updateLedgerInfo(underscore.omit(ledgerInfo, [ '_internal' ]))
-      break
-
-    case appConstants.APP_NAVIGATOR_HANDLER_UNREGISTERED:
-      ledgerInfo.hasBitcoinHandler = false
-      appActions.updateLedgerInfo(underscore.omit(ledgerInfo, [ '_internal' ]))
-      break
-
-    case 'event-set-page-info':
-    case appConstants.APP_WINDOW_BLURRED:
-    case appConstants.APP_CLOSE_WINDOW:
-    case windowConstants.WINDOW_SET_FOCUSED_FRAME:
-    case windowConstants.WINDOW_GOT_RESPONSE_DETAILS:
-      {
-        state = pageDataChanged(state)
-        break
-      }
-  }
-
-  return state
-}
 
 /*
  * module entry points
  */
 
-var init = () => {
-  try {
-    initialize(getSetting(settings.PAYMENTS_ENABLED))
-  } catch (ex) { console.error('ledger.js initialization failed: ', ex) }
-}
-
-var quit = () => {
-  quitP = true
-  visit('NOOP', underscore.now(), null)
-  clearInterval(doneTimer)
-
-  if ((!getSetting(settings.PAYMENTS_ENABLED)) && (getSetting(settings.SHUTDOWN_CLEAR_HISTORY))) reset(true)
-}
 
 var boot = () => {
   if ((bootP) || (client)) return
@@ -337,23 +148,10 @@ var boot = () => {
       bootP = false
       return console.error('ledger client boot error: ', ex)
     }
-    if (client.sync(callback) === true) run(random.randomInt({ min: msecs.minute, max: 10 * msecs.minute }))
+    if (client.sync(callback) === true) run(random.randomInt({ min: miliseconds.minute, max: 10 * miliseconds.minute }))
     getBalance()
 
     bootP = false
-  })
-}
-
-var reset = (doneP) => {
-  var files = [ logPath, publisherPath, scoresPath, synopsisPath ]
-
-  if (!doneP) files.push(statePath)
-  files.forEach((file) => {
-    fs.unlink(pathName(file), (err) => {
-      if ((err) && (err.code !== 'ENOENT')) {
-        console.error('error removing file ' + file + ': ', err)
-      }
-    })
   })
 }
 
@@ -361,159 +159,10 @@ var reset = (doneP) => {
  * Print or Save Recovery Keys
  */
 
-var backupKeys = (appState, action) => {
-  const date = moment().format('L')
-  const paymentId = appState.getIn(['ledgerInfo', 'paymentId'])
-  const passphrase = appState.getIn(['ledgerInfo', 'passphrase'])
-
-  const messageLines = [
-    locale.translation('ledgerBackupText1'),
-    [locale.translation('ledgerBackupText2'), date].join(' '),
-    '',
-    [locale.translation('ledgerBackupText3'), paymentId].join(' '),
-    [locale.translation('ledgerBackupText4'), passphrase].join(' '),
-    '',
-    locale.translation('ledgerBackupText5')
-  ]
-
-  const message = messageLines.join(os.EOL)
-  const filePath = path.join(app.getPath('userData'), '/brave_wallet_recovery.txt')
-
-  fs.writeFile(filePath, message, (err) => {
-    if (err) {
-      console.error(err)
-    } else {
-      tabs.create({url: fileUrl(filePath)}, (webContents) => {
-        if (action.backupAction === 'print') {
-          webContents.print({silent: false, printBackground: false})
-        } else {
-          webContents.downloadURL(fileUrl(filePath), true)
-        }
-      })
-    }
-  })
-
-  return appState
-}
-
-var loadKeysFromBackupFile = (filePath) => {
-  let keys = null
-  let data = fs.readFileSync(filePath)
-
-  if (!data || !data.length || !(data.toString())) {
-    logError('No data in backup file', 'recoveryWallet')
-  } else {
-    try {
-      const recoveryFileContents = data.toString()
-
-      let messageLines = recoveryFileContents.split(os.EOL)
-
-      let paymentIdLine = '' || messageLines[3]
-      let passphraseLine = '' || messageLines[4]
-
-      const paymentIdPattern = new RegExp([locale.translation('ledgerBackupText3'), '([^ ]+)'].join(' '))
-      const paymentId = (paymentIdLine.match(paymentIdPattern) || [])[1]
-
-      const passphrasePattern = new RegExp([locale.translation('ledgerBackupText4'), '(.+)$'].join(' '))
-      const passphrase = (passphraseLine.match(passphrasePattern) || [])[1]
-
-      keys = {
-        paymentId,
-        passphrase
-      }
-    } catch (exc) {
-      logError(exc, 'recoveryWallet')
-    }
-  }
-
-  return keys
-}
 
 /*
  * Recover Ledger Keys
  */
-
-var recoverKeys = (appState, action) => {
-  let firstRecoveryKey, secondRecoveryKey
-
-  if (action.useRecoveryKeyFile) {
-    let recoveryKeyFile = promptForRecoveryKeyFile()
-    if (!recoveryKeyFile) {
-      // user canceled from dialog, we abort without error
-      return appState
-    }
-
-    if (recoveryKeyFile) {
-      let keys = loadKeysFromBackupFile(recoveryKeyFile) || {}
-
-      if (keys) {
-        firstRecoveryKey = keys.paymentId
-        secondRecoveryKey = keys.passphrase
-      }
-    }
-  }
-
-  if (!firstRecoveryKey || !secondRecoveryKey) {
-    firstRecoveryKey = action.firstRecoveryKey
-    secondRecoveryKey = action.secondRecoveryKey
-  }
-
-  const UUID_REGEX = /^[0-9a-z]{8}-[0-9a-z]{4}-[0-9a-z]{4}-[0-9a-z]{4}-[0-9a-z]{12}$/
-  if (typeof firstRecoveryKey !== 'string' || !firstRecoveryKey.match(UUID_REGEX) || typeof secondRecoveryKey !== 'string' || !secondRecoveryKey.match(UUID_REGEX)) {
-    // calling logError sets the error object
-    logError(true, 'recoverKeys')
-    appActions.updateLedgerInfo(underscore.omit(ledgerInfo, [ '_internal' ]))
-    appActions.ledgerRecoveryFailed()
-    return appState
-  }
-
-  client.recoverWallet(firstRecoveryKey, secondRecoveryKey, (err, result) => {
-    let existingLedgerError = ledgerInfo.error
-
-    if (logError(err, 'recoveryWallet')) {
-      // we reset ledgerInfo.error to what it was before (likely null)
-      // if ledgerInfo.error is not null, the wallet info will not display in UI
-      // logError sets ledgerInfo.error, so we must we clear it or UI will show an error
-      ledgerInfo.error = existingLedgerError
-      appActions.updateLedgerInfo(underscore.omit(ledgerInfo, [ '_internal' ]))
-      appActions.ledgerRecoveryFailed()
-    } else {
-      callback(err, result)
-
-      appActions.updateLedgerInfo(underscore.omit(ledgerInfo, [ '_internal' ]))
-      if (balanceTimeoutId) clearTimeout(balanceTimeoutId)
-      getBalance()
-      appActions.ledgerRecoverySucceeded()
-    }
-  })
-
-  return appState
-}
-
-const dialog = electron.dialog
-
-var promptForRecoveryKeyFile = () => {
-  const defaultRecoveryKeyFilePath = path.join(app.getPath('userData'), '/brave_wallet_recovery.txt')
-
-  let files
-
-  if (process.env.SPECTRON) {
-    // skip the dialog for tests
-    console.log(`for test, trying to recover keys from path: ${defaultRecoveryKeyFilePath}`)
-    files = [defaultRecoveryKeyFilePath]
-  } else {
-    files = dialog.showOpenDialog({
-      properties: ['openFile'],
-      defaultPath: defaultRecoveryKeyFilePath,
-      filters: [{
-        name: 'TXT files',
-        extensions: ['txt']
-      }]
-    })
-  }
-
-  return (files && files.length ? files[0] : null)
-}
 
 /*
  * IPC entry point
@@ -570,7 +219,7 @@ if (ipc) {
 
   ipc.on(messages.NOTIFICATION_RESPONSE, (e, message, buttonIndex) => {
     const win = electron.BrowserWindow.getActiveWindow()
-    if (message === addFundsMessage) {
+    if (message === locale.translation('addFundsNotification')) {
       appActions.hideNotification(message)
       // See showNotificationAddFunds() for buttons.
       // buttonIndex === 1 is "Later"; the timestamp until which to delay is set
@@ -584,7 +233,7 @@ if (ipc) {
           windowId: win.id
         })
       }
-    } else if (message === reconciliationMessage) {
+    } else if (message === locale.translation('reconciliationNotification')) {
       appActions.hideNotification(message)
       // buttonIndex === 1 is Dismiss
       if (buttonIndex === 0) {
@@ -600,7 +249,7 @@ if (ipc) {
       if (buttonIndex === 0) {
         appActions.changeSetting(settings.PAYMENTS_NOTIFICATIONS, false)
       }
-    } else if (message === notificationTryPaymentsMessage) {
+    } else if (message === locale.translation('notificationTryPayments')) {
       appActions.hideNotification(message)
       if (buttonIndex === 1 && win) {
         appActions.createTabRequested({
@@ -614,7 +263,7 @@ if (ipc) {
 
   ipc.on(messages.ADD_FUNDS_CLOSED, () => {
     if (balanceTimeoutId) clearTimeout(balanceTimeoutId)
-    balanceTimeoutId = setTimeout(getBalance, 5 * msecs.second)
+    balanceTimeoutId = setTimeout(getBalance, 5 * milisecons.second)
   })
 }
 
@@ -636,266 +285,9 @@ underscore.keys(fileTypes).forEach((fileType) => {
 })
 signatureMax = Math.ceil(signatureMax * 1.5)
 
-const pageDataChanged = (state) => {
-  // NB: in theory we have already seen every element in info except for (perhaps) the last one...
-  const info = pageDataState.getLastInfo(state)
-
-  if (!synopsis || info.isEmpty()) {
-    return
-  }
-
-  if (info.get('url', '').match(/^about/)) {
-    return
-  }
-
-  let publisher = info.get('publisher')
-  const location = info.get('key')
-  if (publisher) {
-    // TODO refactor
-    if (synopsis.publishers[publisher] &&
-      (typeof synopsis.publishers[publisher].faviconURL === 'undefined' || synopsis.publishers[publisher].faviconURL === null)) {
-      getFavIcon(synopsis.publishers[publisher], info, location)
-    }
-
-    // TODO refactor
-    return updateLocation(location, publisher)
-  } else {
-    try {
-      publisher = ledgerPublisher.getPublisher(location, publisherInfo._internal.ruleset.raw)
-      // TODO refactor
-      if (publisher && !blockedP(publisher)) {
-        state = pageDataState.setPublisher(state, location, publisher)
-      } else {
-        publisher = null
-      }
-    } catch (ex) {
-      console.error('getPublisher error for ' + location + ': ' + ex.toString())
-    }
-  }
-
-  if (!publisher) {
-    return
-  }
-
-  const pattern = `https?://${publisher}`
-  const initP = !synopsis.publishers[publisher]
-  // TODO refactor
-  synopsis.initPublisher(publisher)
-
-  if (initP) {
-    // TODO refactor
-    excludeP(publisher, (unused, exclude) => {
-      if (!getSetting(settings.PAYMENTS_SITES_AUTO_SUGGEST)) {
-        exclude = false
-      } else {
-        exclude = !exclude
-      }
-      appActions.changeSiteSetting(pattern, 'ledgerPayments', exclude)
-      updatePublisherInfo()
-    })
-  }
-  // TODO refactor
-  updateLocation(location, publisher)
-  // TODO refactor
-  getFavIcon(synopsis.publishers[publisher], info, location)
-
-  const pageLoad = pageDataState.getLoad(state)
-  const view = pageDataState.getView(state)
-
-  if (ledgerUtil.shouldTrackView(view, pageLoad)) {
-    // TODO refactor
-    visit(view.get('url', 'NOOP'), view.get('timestamp', underscore.now()), view.get('tabId'))
-  }
-
-  return state
-}
-
 /*
  * module initialization
  */
-
-var initialize = (paymentsEnabled) => {
-  var ruleset
-
-  if (!v2RulesetDB) v2RulesetDB = levelup(pathName(v2RulesetPath))
-  if (!v2PublishersDB) v2PublishersDB = levelup(pathName(v2PublishersPath))
-  enable(paymentsEnabled)
-
-  // Check if relevant browser notifications should be shown every 15 minutes
-  if (notificationTimeout) clearInterval(notificationTimeout)
-  notificationTimeout = setInterval(showNotifications, 15 * msecs.minute)
-
-  if (!paymentsEnabled) {
-    client = null
-    return appActions.updateLedgerInfo({})
-  }
-  if (client) return
-
-  if (!ledgerPublisher) ledgerPublisher = require('ledger-publisher')
-  ruleset = []
-  ledgerPublisher.ruleset.forEach(rule => { if (rule.consequent) ruleset.push(rule) })
-  cacheRuleSet(ruleset)
-
-  fs.access(pathName(statePath), fs.FF_OK, (err) => {
-    if (!err) {
-      if (clientOptions.verboseP) console.log('\nfound ' + pathName(statePath))
-
-      fs.readFile(pathName(statePath), (err, data) => {
-        var state
-
-        if (err) return console.error('read error: ' + err.toString())
-
-        try {
-          state = JSON.parse(data)
-          if (clientOptions.verboseP) console.log('\nstarting up ledger client integration')
-        } catch (ex) {
-          return console.error('statePath parse error: ' + ex.toString())
-        }
-
-        getStateInfo(state)
-
-        try {
-          var timeUntilReconcile
-          clientprep()
-          client = ledgerClient(state.personaId,
-                                underscore.extend(state.options, { roundtrip: roundtrip }, clientOptions),
-                                state)
-
-          // Scenario: User enables Payments, disables it, waits 30+ days, then
-          // enables it again -> reconcileStamp is in the past.
-          // In this case reset reconcileStamp to the future.
-          try { timeUntilReconcile = client.timeUntilReconcile() } catch (ex) {}
-          let ledgerWindow = (synopsis.options.numFrames - 1) * synopsis.options.frameSize
-          if (typeof timeUntilReconcile === 'number' && timeUntilReconcile < -ledgerWindow) {
-            client.setTimeUntilReconcile(null, (err, stateResult) => {
-              if (err) return console.error('ledger setTimeUntilReconcile error: ' + err.toString())
-
-              if (!stateResult) {
-                return
-              }
-              getStateInfo(stateResult)
-
-              muonWriter(pathName(statePath), stateResult)
-            })
-          }
-        } catch (ex) {
-          return console.error('ledger client creation error: ', ex)
-        }
-
-        // speed-up browser start-up by delaying the first synchronization action
-        setTimeout(() => {
-          if (!client) return
-
-          if (client.sync(callback) === true) run(random.randomInt({ min: msecs.minute, max: 10 * msecs.minute }))
-          cacheRuleSet(state.ruleset)
-        }, 3 * msecs.second)
-
-        // Make sure bravery props are up-to-date with user settings
-        if (!ledgerInfo.address) ledgerInfo.address = client.getWalletAddress()
-        setPaymentInfo(getSetting(settings.PAYMENTS_CONTRIBUTION_AMOUNT))
-        getBalance()
-      })
-      return
-    }
-
-    if (err.code !== 'ENOENT') console.error('statePath read error: ' + err.toString())
-    appActions.updateLedgerInfo({})
-  })
-}
-
-var clientprep = () => {
-  if (!ledgerClient) ledgerClient = require('ledger-client')
-  ledgerInfo._internal.debugP = ledgerClient.prototype.boolion(process.env.LEDGER_CLIENT_DEBUG)
-  publisherInfo._internal.debugP = ledgerClient.prototype.boolion(process.env.LEDGER_PUBLISHER_DEBUG)
-  publisherInfo._internal.verboseP = ledgerClient.prototype.boolion(process.env.LEDGER_PUBLISHER_VERBOSE)
-}
-
-var enable = (paymentsEnabled) => {
-  if (paymentsEnabled && !getSetting(settings.PAYMENTS_NOTIFICATION_TRY_PAYMENTS_DISMISSED)) {
-    appActions.changeSetting(settings.PAYMENTS_NOTIFICATION_TRY_PAYMENTS_DISMISSED, true)
-  }
-
-  publisherInfo._internal.enabled = paymentsEnabled
-  if (synopsis) return updatePublisherInfo()
-
-  if (!ledgerPublisher) ledgerPublisher = require('ledger-publisher')
-  synopsis = new (ledgerPublisher.Synopsis)()
-  fs.readFile(pathName(synopsisPath), (err, data) => {
-    var initSynopsis = () => {
-      var value
-
-      // cf., the `Synopsis` constructor, https://github.com/brave/ledger-publisher/blob/master/index.js#L167
-      value = getSetting(settings.PAYMENTS_MINIMUM_VISIT_TIME)
-      if (!value) {
-        value = 8 * 1000
-        appActions.changeSetting(settings.PAYMENTS_MINIMUM_VISIT_TIME, value)
-      }
-
-      // for earlier versions of the code...
-      if ((value > 0) && (value < 1000)) synopsis.options.minPublisherDuration = value * 1000
-
-      value = getSetting(settings.PAYMENTS_MINIMUM_VISITS)
-      if (!value) {
-        value = 1
-        appActions.changeSetting(settings.PAYMENTS_MINIMUM_VISITS, value)
-      }
-      if (value > 0) synopsis.options.minPublisherVisits = value
-
-      if (process.env.NODE_ENV === 'test') {
-        synopsis.options.minPublisherDuration = 0
-        synopsis.options.minPublisherVisits = 0
-      } else {
-        if (process.env.LEDGER_PUBLISHER_MIN_DURATION) {
-          synopsis.options.minPublisherDuration = ledgerClient.prototype.numbion(process.env.LEDGER_PUBLISHER_MIN_DURATION)
-        }
-        if (process.env.LEDGER_PUBLISHER_MIN_VISITS) {
-          synopsis.options.minPublisherVisits = ledgerClient.prototype.numbion(process.env.LEDGER_PUBLISHER_MIN_VISITS)
-        }
-      }
-
-      underscore.keys(synopsis.publishers).forEach((publisher) => {
-        excludeP(publisher)
-        verifiedP(publisher)
-      })
-
-      updatePublisherInfo()
-    }
-
-    if (publisherInfo._internal.verboseP) console.log('\nstarting up ledger publisher integration')
-
-    if (err) {
-      if (err.code !== 'ENOENT') console.error('synopsisPath read error: ' + err.toString())
-      initSynopsis()
-      return updatePublisherInfo()
-    }
-
-    if (publisherInfo._internal.verboseP) console.log('\nfound ' + pathName(synopsisPath))
-    try {
-      synopsis = new (ledgerPublisher.Synopsis)(data)
-    } catch (ex) {
-      console.error('synopsisPath parse error: ' + ex.toString())
-    }
-    initSynopsis()
-    underscore.keys(synopsis.publishers).forEach((publisher) => {
-      if (synopsis.publishers[publisher].faviconURL === null) delete synopsis.publishers[publisher].faviconURL
-    })
-    updatePublisherInfo()
-
-    // change undefined include publishers to include publishers
-    appActions.enableUndefinedPublishers(synopsis.publishers)
-
-    fs.unlink(pathName(publisherPath), (err) => {
-      if ((err) && (err.code !== 'ENOENT')) {
-        console.error('error removing file ' + pathName(publisherPath) + ': ', err)
-      }
-    })
-    fs.unlink(pathName(scoresPath), (err) => {
-      if ((err) && (err.code !== 'ENOENT')) {
-        console.error('error removing file ' + pathName(scoresPath) + ': ', err)
-      }
-    })
-  })
-}
 
 /*
  * update location information
@@ -938,39 +330,23 @@ var updateLocation = (location, publisher) => {
   if (updateP) updateLocationInfo(location)
 }
 
-/*
- * update publisher information
- */
-
-var publisherInfo = {
-  options: undefined,
-
-  synopsis: undefined,
-
-  _internal: {
-    enabled: false,
-
-    ruleset: { raw: [], cooked: [] }
-  }
-}
-
-const getFavIcon = (entry, page, location) => {
-  if ((page.protocol) && (!entry.protocol)) {
-    entry.protocol = page.protocol
+const getFavIcon = (publisher, page, location) => {
+  if ((page.protocol) && (!publisher.protocol)) {
+    publisher.protocol = page.protocol
   }
 
-  if ((typeof entry.faviconURL === 'undefined') && ((page.faviconURL) || (entry.protocol))) {
-    let faviconURL = page.faviconURL || entry.protocol + '//' + urlParse(location).host + '/favicon.ico'
+  if ((typeof publisher.faviconURL === 'undefined') && ((page.faviconURL) || (publisher.protocol))) {
+    let faviconURL = page.faviconURL || publisher.protocol + '//' + urlParse(location).host + '/favicon.ico'
     if (publisherInfo._internal.debugP) {
       console.log('\nrequest: ' + faviconURL)
     }
 
-    entry.faviconURL = null
-    fetchFavIcon(entry, faviconURL)
+    publisher.faviconURL = null
+    fetchFavIcon(publisher, faviconURL)
   }
 }
 
-const fetchFavIcon = (entry, url, redirects) => {
+const fetchFavIcon = (publisher, url, redirects) => {
   if (typeof redirects === 'undefined') redirects = 0
 
   request.request({ url: url, responseType: 'blob' }, (err, response, blob) => {
@@ -997,7 +373,7 @@ const fetchFavIcon = (entry, url, redirects) => {
     }
 
     if ((response.statusCode === 301) && (response.headers.location)) {
-      if (redirects < 3) fetchFavIcon(entry, response.headers.location, redirects++)
+      if (redirects < 3) fetchFavIcon(publisher, response.headers.location, redirects++)
       return null
     }
 
@@ -1027,500 +403,20 @@ const fetchFavIcon = (entry, url, redirects) => {
     } else if ((tail > 0) && (tail + 8 >= blob.length)) return
 
     if (publisherInfo._internal.debugP) {
-      console.log('\n' + entry.site + ' synopsis=' +
-        JSON.stringify(underscore.extend(underscore.omit(entry, [ 'faviconURL', 'window' ]),
-          { faviconURL: entry.faviconURL && '... ' }), null, 2))
+      console.log('\n' + publisher.site + ' synopsis=' +
+        JSON.stringify(underscore.extend(underscore.omit(publisher, [ 'faviconURL', 'window' ]),
+          { faviconURL: publisher.faviconURL && '... ' }), null, 2))
     }
 
-    entry.faviconURL = blob
+    publisher.faviconURL = blob
     updatePublisherInfo()
   })
 }
 
-var updatePublisherInfo = (changedPublisher) => {
-  var data
-
-  muonWriter(pathName(synopsisPath), synopsis)
-  if (!publisherInfo._internal.enabled) return
-
-  publisherInfo.synopsis = synopsisNormalizer(changedPublisher)
-  publisherInfo.synopsisOptions = synopsis.options
-
-  if (publisherInfo._internal.debugP) {
-    data = []
-    publisherInfo.synopsis.forEach((entry) => {
-      data.push(underscore.extend(underscore.omit(entry, [ 'faviconURL' ]), { faviconURL: entry.faviconURL && '...' }))
-    })
-
-    console.log('\nupdatePublisherInfo: ' + JSON.stringify({ options: publisherInfo.synopsisOptions, synopsis: data }, null, 2))
-  }
-
-  appActions.updatePublisherInfo(underscore.omit(publisherInfo, [ '_internal' ]))
-}
-
-var blockedP = (publisher) => {
-  var siteSetting = appStore.getState().get('siteSettings').get(`https?://${publisher}`)
-
-  return ((!!siteSetting) && (siteSetting.get('ledgerPaymentsShown') === false))
-}
-
-var stickyP = (publisher) => {
-  var siteSettings = appStore.getState().get('siteSettings')
-  var pattern = `https?://${publisher}`
-  var siteSetting = siteSettings.get(pattern)
-  var result = (siteSetting) && (siteSetting.get('ledgerPayments'))
-
-  // NB: legacy clean-up
-  if ((typeof result === 'undefined') && (typeof synopsis.publishers[publisher].options.stickyP !== 'undefined')) {
-    result = synopsis.publishers[publisher].options.stickyP
-    appActions.changeSiteSetting(pattern, 'ledgerPayments', result)
-  }
-
-  if (synopsis.publishers[publisher] &&
-    synopsis.publishers[publisher].options &&
-    synopsis.publishers[publisher].options.stickyP) {
-    delete synopsis.publishers[publisher].options.stickyP
-  }
-
-  return (result === undefined || result)
-}
-
-var eligibleP = (publisher) => {
-  if (!synopsis.options.minPublisherDuration && process.env.NODE_ENV !== 'test') {
-    synopsis.options.minPublisherDuration = getSetting(settings.PAYMENTS_MINIMUM_VISIT_TIME)
-  }
-
-  return ((synopsis.publishers[publisher].scores[synopsis.options.scorekeeper] > 0) &&
-          (synopsis.publishers[publisher].duration >= synopsis.options.minPublisherDuration) &&
-          (synopsis.publishers[publisher].visits >= synopsis.options.minPublisherVisits))
-}
-
-var visibleP = (publisher) => {
-  if (synopsis.options.showOnlyVerified === undefined) {
-    synopsis.options.showOnlyVerified = getSetting(settings.PAYMENTS_ALLOW_NON_VERIFIED)
-  }
-
-  const publisherOptions = synopsis.publishers[publisher].options
-  const onlyVerified = !synopsis.options.showOnlyVerified
-
-  // Publisher Options
-  const excludedByUser = blockedP(publisher)
-  const eligibleByPublisherToggle = stickyP(publisher) != null
-  const eligibleByNumberOfVisits = eligibleP(publisher)
-  const isInExclusionList = publisherOptions && publisherOptions.exclude
-  const verifiedPublisher = publisherOptions && publisherOptions.verified
-
-  // websites not included in exclusion list are eligible by number of visits
-  // but can be enabled by user action in the publisher toggle
-  const isEligible = (eligibleByNumberOfVisits && !isInExclusionList) || eligibleByPublisherToggle
-
-  // If user decide to remove the website, don't show it.
-  if (excludedByUser) {
-    return false
-  }
-
-  // Unless user decided to enable publisher with publisherToggle,
-  // do not show exclusion list.
-  if (!eligibleByPublisherToggle && isInExclusionList) {
-    return false
-  }
-
-  // If verified option is set, only show verified publishers
-  if (isEligible && onlyVerified) {
-    return verifiedPublisher
-  }
-
-  return isEligible
-}
-
-var contributeP = (publisher) => {
-  return (((stickyP(publisher)) || (synopsis.publishers[publisher].options.exclude !== true)) &&
-          (eligibleP(publisher)) &&
-          (!blockedP(publisher)))
-}
-
-var synopsisNormalizer = (changedPublisher) => {
-  // courtesy of https://stackoverflow.com/questions/13483430/how-to-make-rounded-percentages-add-up-to-100#13485888
-  const roundToTarget = (l, target, property) => {
-    let off = target - underscore.reduce(l, (acc, x) => { return acc + Math.round(x[property]) }, 0)
-
-    return underscore.sortBy(l, (x) => Math.round(x[property]) - x[property])
-      .map((x, i) => {
-        x[property] = Math.round(x[property]) + (off > i) - (i >= (l.length + off))
-        return x
-      })
-  }
-
-  const normalizePinned = (dataPinned, total, target, setOne) => dataPinned.map((publisher) => {
-    let newPer
-    let floatNumber
-
-    if (setOne) {
-      newPer = 1
-      floatNumber = 1
-    } else {
-      floatNumber = (publisher.pinPercentage / total) * target
-      newPer = Math.floor(floatNumber)
-      if (newPer < 1) {
-        newPer = 1
-      }
-    }
-
-    publisher.weight = floatNumber
-    publisher.pinPercentage = newPer
-    return publisher
-  })
-
-  const getPublisherData = (result) => {
-    let duration = result.duration
-
-    let data = {
-      verified: result.options.verified || false,
-      site: result.publisher,
-      views: result.visits,
-      duration: duration,
-      daysSpent: 0,
-      hoursSpent: 0,
-      minutesSpent: 0,
-      secondsSpent: 0,
-      faviconURL: result.faviconURL,
-      score: result.scores[scorekeeper],
-      pinPercentage: result.pinPercentage,
-      weight: result.pinPercentage
-    }
-    // HACK: Protocol is sometimes blank here, so default to http:// so we can
-    // still generate publisherURL.
-    data.publisherURL = (result.protocol || 'http:') + '//' + result.publisher
-
-    if (duration >= msecs.day) {
-      data.daysSpent = Math.max(Math.round(duration / msecs.day), 1)
-    } else if (duration >= msecs.hour) {
-      data.hoursSpent = Math.max(Math.floor(duration / msecs.hour), 1)
-      data.minutesSpent = Math.round((duration % msecs.hour) / msecs.minute)
-    } else if (duration >= msecs.minute) {
-      data.minutesSpent = Math.max(Math.round(duration / msecs.minute), 1)
-      data.secondsSpent = Math.round((duration % msecs.minute) / msecs.second)
-    } else {
-      data.secondsSpent = Math.max(Math.round(duration / msecs.second), 1)
-    }
-
-    return data
-  }
-
-  let results
-  let dataPinned = []
-  let dataUnPinned = []
-  let dataExcluded = []
-  let pinnedTotal = 0
-  let unPinnedTotal = 0
-  const scorekeeper = synopsis.options.scorekeeper
-
-  results = []
-  underscore.keys(synopsis.publishers).forEach((publisher) => {
-    if (!visibleP(publisher)) return
-
-    results.push(underscore.extend({publisher: publisher}, underscore.omit(synopsis.publishers[publisher], 'window')))
-  }, synopsis)
-  results = underscore.sortBy(results, (entry) => { return -entry.scores[scorekeeper] })
-
-  // move publisher to the correct array and get totals
-  results.forEach((result) => {
-    if (result.pinPercentage && result.pinPercentage > 0) {
-      // pinned
-      pinnedTotal += result.pinPercentage
-      dataPinned.push(getPublisherData(result))
-    } else if (stickyP(result.publisher)) {
-      // unpinned
-      unPinnedTotal += result.scores[scorekeeper]
-      dataUnPinned.push(result)
-    } else {
-      // excluded
-      let publisher = getPublisherData(result)
-      publisher.percentage = 0
-      publisher.weight = 0
-      dataExcluded.push(publisher)
-    }
-  })
-
-  // round if over 100% of pinned publishers
-  if (pinnedTotal > 100) {
-    if (changedPublisher) {
-      const changedObject = dataPinned.filter(publisher => publisher.site === changedPublisher)[0]
-      const setOne = changedObject.pinPercentage > (100 - dataPinned.length - 1)
-
-      if (setOne) {
-        changedObject.pinPercentage = 100 - dataPinned.length + 1
-        changedObject.weight = changedObject.pinPercentage
-      }
-
-      const pinnedRestTotal = pinnedTotal - changedObject.pinPercentage
-      dataPinned = dataPinned.filter(publisher => publisher.site !== changedPublisher)
-      dataPinned = normalizePinned(dataPinned, pinnedRestTotal, (100 - changedObject.pinPercentage), setOne)
-      dataPinned = roundToTarget(dataPinned, (100 - changedObject.pinPercentage), 'pinPercentage')
-
-      dataPinned.push(changedObject)
-    } else {
-      dataPinned = normalizePinned(dataPinned, pinnedTotal, 100)
-      dataPinned = roundToTarget(dataPinned, 100, 'pinPercentage')
-    }
-
-    dataUnPinned = dataUnPinned.map((result) => {
-      let publisher = getPublisherData(result)
-      publisher.percentage = 0
-      publisher.weight = 0
-      return publisher
-    })
-
-    // sync app store
-    appActions.changeLedgerPinnedPercentages(dataPinned)
-  } else if (dataUnPinned.length === 0 && pinnedTotal < 100) {
-    // when you don't have any unpinned sites and pinned total is less then 100 %
-    dataPinned = normalizePinned(dataPinned, pinnedTotal, 100, false)
-    dataPinned = roundToTarget(dataPinned, 100, 'pinPercentage')
-
-    // sync app store
-    appActions.changeLedgerPinnedPercentages(dataPinned)
-  } else {
-    // unpinned publishers
-    dataUnPinned = dataUnPinned.map((result) => {
-      let publisher = getPublisherData(result)
-      const floatNumber = (publisher.score / unPinnedTotal) * (100 - pinnedTotal)
-      publisher.percentage = Math.round(floatNumber)
-      publisher.weight = floatNumber
-      return publisher
-    })
-
-    // normalize unpinned values
-    dataUnPinned = roundToTarget(dataUnPinned, (100 - pinnedTotal), 'percentage')
-  }
-
-  const newData = dataPinned.concat(dataUnPinned, dataExcluded)
-
-  // sync synopsis
-  newData.forEach((item) => {
-    synopsis.publishers[item.site].weight = item.weight
-    synopsis.publishers[item.site].pinPercentage = item.pinPercentage
-  })
-
-  return newData
-}
 
 /*
  * publisher utilities
  */
-
-var currentLocation = 'NOOP'
-var currentTimestamp = underscore.now()
-
-var visit = (location, timestamp, tabId) => {
-  var setLocation = () => {
-    var duration, publisher, revisitP
-
-    if (!synopsis) return
-
-    if (publisherInfo._internal.verboseP) {
-      console.log('locations[' + currentLocation + ']=' + JSON.stringify(locations[currentLocation], null, 2) +
-                  ' duration=' + (timestamp - currentTimestamp) + ' msec' + ' tabId=' + tabId)
-    }
-    if ((location === currentLocation) || (!locations[currentLocation]) || (!tabId)) return
-
-    publisher = locations[currentLocation].publisher
-    if (!publisher) return
-
-    if (!publishers[publisher]) publishers[publisher] = {}
-    if (!publishers[publisher][currentLocation]) publishers[publisher][currentLocation] = { tabIds: [] }
-    publishers[publisher][currentLocation].timestamp = timestamp
-    revisitP = publishers[publisher][currentLocation].tabIds.indexOf(tabId) !== -1
-    if (!revisitP) publishers[publisher][currentLocation].tabIds.push(tabId)
-
-    duration = timestamp - currentTimestamp
-    if (publisherInfo._internal.verboseP) {
-      console.log('\nadd publisher ' + publisher + ': ' + duration + ' msec' + ' revisitP=' + revisitP + ' state=' +
-                  JSON.stringify(underscore.extend({ location: currentLocation }, publishers[publisher][currentLocation]),
-                                 null, 2))
-    }
-    synopsis.addPublisher(publisher, { duration: duration, revisitP: revisitP })
-    updatePublisherInfo()
-    verifiedP(publisher)
-  }
-
-  setLocation()
-  if (location === currentLocation) return
-
-  currentLocation = location.match(/^about/) ? 'NOOP' : location
-  currentTimestamp = timestamp
-}
-
-var cacheRuleSet = (ruleset) => {
-  var stewed, syncP
-
-  if ((!ruleset) || (underscore.isEqual(publisherInfo._internal.ruleset.raw, ruleset))) return
-
-  try {
-    stewed = []
-    ruleset.forEach((rule) => {
-      var entry = { condition: acorn.parse(rule.condition) }
-
-      if (rule.dom) {
-        if (rule.dom.publisher) {
-          entry.publisher = { selector: rule.dom.publisher.nodeSelector,
-            consequent: acorn.parse(rule.dom.publisher.consequent)
-          }
-        }
-        if (rule.dom.faviconURL) {
-          entry.faviconURL = { selector: rule.dom.faviconURL.nodeSelector,
-            consequent: acorn.parse(rule.dom.faviconURL.consequent)
-          }
-        }
-      }
-      if (!entry.publisher) entry.consequent = rule.consequent ? acorn.parse(rule.consequent) : rule.consequent
-
-      stewed.push(entry)
-    })
-
-    publisherInfo._internal.ruleset.raw = ruleset
-    publisherInfo._internal.ruleset.cooked = stewed
-    if (!synopsis) return
-
-    underscore.keys(synopsis.publishers).forEach((publisher) => {
-      var location = (synopsis.publishers[publisher].protocol || 'http:') + '//' + publisher
-      var ctx = urlParse(location, true)
-
-      ctx.TLD = tldjs.getPublicSuffix(ctx.host)
-      if (!ctx.TLD) return
-
-      ctx = underscore.mapObject(ctx, function (value, key) { if (!underscore.isFunction(value)) return value })
-      ctx.URL = location
-      ctx.SLD = tldjs.getDomain(ctx.host)
-      ctx.RLD = tldjs.getSubdomain(ctx.host)
-      ctx.QLD = ctx.RLD ? underscore.last(ctx.RLD.split('.')) : ''
-
-      stewed.forEach((rule) => {
-        if ((rule.consequent !== null) || (rule.dom)) return
-        if (!rulesolver.resolve(rule.condition, ctx)) return
-
-        if (publisherInfo._internal.verboseP) console.log('\npurging ' + publisher)
-        delete synopsis.publishers[publisher]
-        delete publishers[publisher]
-        syncP = true
-      })
-    })
-    if (!syncP) return
-
-    updatePublisherInfo()
-  } catch (ex) {
-    console.error('ruleset error: ', ex)
-  }
-}
-
-var excludeP = (publisher, callback) => {
-  var doneP
-
-  var done = (err, result) => {
-    doneP = true
-    if ((!err) && (typeof result !== 'undefined') && (!!synopsis.publishers[publisher]) &&
-        (synopsis.publishers[publisher].options.exclude !== result)) {
-      synopsis.publishers[publisher].options.exclude = result
-      updatePublisherInfo()
-    }
-
-    if (callback) callback(err, result)
-  }
-
-  if (!v2RulesetDB) return setTimeout(() => { excludeP(publisher, callback) }, 5 * msecs.second)
-
-  inspectP(v2RulesetDB, v2RulesetPath, publisher, 'exclude', 'domain:' + publisher, (err, result) => {
-    var props
-
-    if (!err) return done(err, result.exclude)
-
-    props = ledgerPublisher.getPublisherProps('https://' + publisher)
-    if (!props) return done()
-
-    v2RulesetDB.createReadStream({ lt: 'domain:' }).on('data', (data) => {
-      var regexp, result, sldP, tldP
-
-      if (doneP) return
-
-      sldP = data.key.indexOf('SLD:') === 0
-      tldP = data.key.indexOf('TLD:') === 0
-      if ((!tldP) && (!sldP)) return
-
-      if (underscore.intersection(data.key.split(''),
-                                   [ '^', '$', '*', '+', '?', '[', '(', '{', '|' ]).length === 0) {
-        if ((data.key !== ('TLD:' + props.TLD)) && (props.SLD && data.key !== ('SLD:' + props.SLD.split('.')[0]))) return
-      } else {
-        try {
-          regexp = new RegExp(data.key.substr(4))
-          if (!regexp.test(props[tldP ? 'TLD' : 'SLD'])) return
-        } catch (ex) {
-          console.error(v2RulesetPath + ' stream invalid regexp ' + data.key + ': ' + ex.toString())
-        }
-      }
-
-      try {
-        result = JSON.parse(data.value)
-      } catch (ex) {
-        console.error(v2RulesetPath + ' stream invalid JSON ' + data.entry + ': ' + data.value)
-      }
-
-      done(null, result.exclude)
-    }).on('error', (err) => {
-      console.error(v2RulesetPath + ' stream error: ' + JSON.stringify(err, null, 2))
-    }).on('close', () => {
-    }).on('end', () => {
-      if (!doneP) done(null, false)
-    })
-  })
-}
-
-var verifiedP = (publisher, callback) => {
-  inspectP(v2PublishersDB, v2PublishersPath, publisher, 'verified', null, callback)
-
-  if (process.env.NODE_ENV === 'test') {
-    testVerifiedPublishers.forEach((publisher) => {
-      if (synopsis.publishers[publisher]) {
-        if (!synopsis.publishers[publisher].options) {
-          synopsis.publishers[publisher].options = {}
-        }
-
-        synopsis.publishers[publisher].options.verified = true
-      }
-    })
-    updatePublisherInfo()
-  }
-}
-
-var inspectP = (db, path, publisher, property, key, callback) => {
-  var done = (err, result) => {
-    if ((!err) && (typeof result !== 'undefined') && (!!synopsis.publishers[publisher]) &&
-        (synopsis.publishers[publisher].options[property] !== result[property])) {
-      synopsis.publishers[publisher].options[property] = result[property]
-      updatePublisherInfo()
-    }
-
-    if (callback) callback(err, result)
-  }
-
-  if (!key) key = publisher
-  db.get(key, (err, value) => {
-    var result
-
-    if (err) {
-      if (!err.notFound) console.error(path + ' get ' + key + ' error: ' + JSON.stringify(err, null, 2))
-      return done(err)
-    }
-
-    try {
-      result = JSON.parse(value)
-    } catch (ex) {
-      console.error(v2RulesetPath + ' stream invalid JSON ' + key + ': ' + value)
-      result = {}
-    }
-
-    done(null, result)
-  })
-}
 
 /*
  * update ledger information
@@ -1603,686 +499,15 @@ var ledgerInfo = {
   error: null
 }
 
-var updateLedgerInfo = () => {
-  var info = ledgerInfo._internal.paymentInfo
-  var now = underscore.now()
-
-  if (info) {
-    underscore.extend(ledgerInfo,
-                      underscore.pick(info, [ 'address', 'passphrase', 'balance', 'unconfirmed', 'satoshis', 'btc', 'amount',
-                        'currency' ]))
-    if ((!info.buyURLExpires) || (info.buyURLExpires > now)) {
-      ledgerInfo.buyURL = info.buyURL
-      ledgerInfo.buyMaximumUSD = 6
-    }
-    if (typeof process.env.ADDFUNDS_URL !== 'undefined') {
-      ledgerInfo.buyURLFrame = true
-      ledgerInfo.buyURL = process.env.ADDFUNDS_URL + '?' +
-                          querystring.stringify({ currency: ledgerInfo.currency,
-                            amount: getSetting(settings.PAYMENTS_CONTRIBUTION_AMOUNT),
-                            address: ledgerInfo.address })
-      ledgerInfo.buyMaximumUSD = false
-    }
-
-    underscore.extend(ledgerInfo, ledgerInfo._internal.cache || {})
-  }
-
-  if ((client) && (now > ledgerInfo._internal.geoipExpiry)) {
-    ledgerInfo._internal.geoipExpiry = now + (5 * msecs.minute)
-
-    if (!ledgerGeoIP) ledgerGeoIP = require('ledger-geoip')
-    return ledgerGeoIP.getGeoIP(client.options, (err, provider, result) => {
-      if (err) console.warn('ledger geoip warning: ' + JSON.stringify(err, null, 2))
-      if (result) ledgerInfo.countryCode = result
-
-      ledgerInfo.exchangeInfo = ledgerInfo._internal.exchanges[ledgerInfo.countryCode]
-
-      if (now <= ledgerInfo._internal.exchangeExpiry) return updateLedgerInfo()
-
-      ledgerInfo._internal.exchangeExpiry = now + msecs.day
-      roundtrip({ path: '/v1/exchange/providers' }, client.options, (err, response, body) => {
-        if (err) console.error('ledger exchange error: ' + JSON.stringify(err, null, 2))
-
-        ledgerInfo._internal.exchanges = body || {}
-        ledgerInfo.exchangeInfo = ledgerInfo._internal.exchanges[ledgerInfo.countryCode]
-        updateLedgerInfo()
-      })
-    })
-  }
-
-  if (ledgerInfo._internal.debugP) {
-    console.log('\nupdateLedgerInfo: ' + JSON.stringify(underscore.omit(ledgerInfo, [ '_internal' ]), null, 2))
-  }
-
-  appActions.updateLedgerInfo(underscore.omit(ledgerInfo, [ '_internal' ]))
-}
-
 /*
  * ledger client callbacks
  */
-
-var callback = (err, result, delayTime) => {
-  var results
-  var entries = client && client.report()
-
-  if (clientOptions.verboseP) {
-    console.log('\nledger client callback: clientP=' + (!!client) + ' errP=' + (!!err) + ' resultP=' + (!!result) +
-                ' delayTime=' + delayTime)
-  }
-
-  if (err) {
-    console.log('ledger client error(1): ' + JSON.stringify(err, null, 2) + (err.stack ? ('\n' + err.stack) : ''))
-    if (!client) return
-
-    if (typeof delayTime === 'undefined') delayTime = random.randomInt({ min: msecs.minute, max: 10 * msecs.minute })
-  }
-
-  if (!result) return run(delayTime)
-
-  if ((client) && (result.properties.wallet)) {
-    if (!ledgerInfo.created) setPaymentInfo(getSetting(settings.PAYMENTS_CONTRIBUTION_AMOUNT))
-
-    getStateInfo(result)
-    getPaymentInfo()
-  }
-  cacheRuleSet(result.ruleset)
-  if (result.rulesetV2) {
-    results = result.rulesetV2
-    delete result.rulesetV2
-
-    entries = []
-    results.forEach((entry) => {
-      var key = entry.facet + ':' + entry.publisher
-
-      if (entry.exclude !== false) {
-        entries.push({ type: 'put', key: key, value: JSON.stringify(underscore.omit(entry, [ 'facet', 'publisher' ])) })
-      } else {
-        entries.push({ type: 'del', key: key })
-      }
-    })
-
-    v2RulesetDB.batch(entries, (err) => {
-      if (err) return console.error(v2RulesetPath + ' error: ' + JSON.stringify(err, null, 2))
-
-      if (entries.length === 0) return
-
-      underscore.keys(synopsis.publishers).forEach((publisher) => {
-// be safe...
-        if (synopsis.publishers[publisher]) delete synopsis.publishers[publisher].options.exclude
-
-        excludeP(publisher)
-      })
-    })
-  }
-  if (result.publishersV2) {
-    results = result.publishersV2
-    delete result.publishersV2
-
-    entries = []
-    results.forEach((entry) => {
-      entries.push({ type: 'put',
-        key: entry.publisher,
-        value: JSON.stringify(underscore.omit(entry, [ 'publisher' ]))
-      })
-      if ((synopsis.publishers[entry.publisher]) &&
-          (synopsis.publishers[entry.publisher].options.verified !== entry.verified)) {
-        synopsis.publishers[entry.publisher].options.verified = entry.verified
-        updatePublisherInfo()
-      }
-    })
-    v2PublishersDB.batch(entries, (err) => {
-      if (err) return console.error(v2PublishersPath + ' error: ' + JSON.stringify(err, null, 2))
-    })
-  }
-
-  muonWriter(pathName(statePath), result)
-  run(delayTime)
-}
-
-var roundtrip = (params, options, callback) => {
-  var i
-  var parts = typeof params.server === 'string' ? urlParse(params.server)
-                : typeof params.server !== 'undefined' ? params.server
-                : typeof options.server === 'string' ? urlParse(options.server) : options.server
-  var rawP = options.rawP
-
-  if (!params.method) params.method = 'GET'
-  parts = underscore.extend(underscore.pick(parts, [ 'protocol', 'hostname', 'port' ]),
-                            underscore.omit(params, [ 'headers', 'payload', 'timeout' ]))
-
-// TBD: let the user configure this via preferences [MTR]
-  if ((parts.hostname === 'ledger.brave.com') && (params.useProxy)) parts.hostname = 'ledger-proxy.privateinternetaccess.com'
-
-  i = parts.path.indexOf('?')
-  if (i !== -1) {
-    parts.pathname = parts.path.substring(0, i)
-    parts.search = parts.path.substring(i)
-  } else {
-    parts.pathname = parts.path
-  }
-
-  options = {
-    url: urlFormat(parts),
-    method: params.method,
-    payload: params.payload,
-    responseType: 'text',
-    headers: underscore.defaults(params.headers || {}, { 'content-type': 'application/json; charset=utf-8' }),
-    verboseP: options.verboseP
-  }
-  request.request(options, (err, response, body) => {
-    var payload
-
-    if ((response) && (options.verboseP)) {
-      console.log('[ response for ' + params.method + ' ' + parts.protocol + '//' + parts.hostname + params.path + ' ]')
-      console.log('>>> HTTP/' + response.httpVersionMajor + '.' + response.httpVersionMinor + ' ' + response.statusCode +
-                 ' ' + (response.statusMessage || ''))
-      underscore.keys(response.headers).forEach((header) => { console.log('>>> ' + header + ': ' + response.headers[header]) })
-      console.log('>>>')
-      console.log('>>> ' + (body || '').split('\n').join('\n>>> '))
-    }
-
-    if (err) return callback(err)
-
-    if (Math.floor(response.statusCode / 100) !== 2) {
-      return callback(new Error('HTTP response ' + response.statusCode) + ' for ' + params.method + ' ' + params.path)
-    }
-
-    try {
-      payload = rawP ? body : (response.statusCode !== 204) ? JSON.parse(body) : null
-    } catch (err) {
-      return callback(err)
-    }
-
-    try {
-      callback(null, response, payload)
-    } catch (err0) {
-      if (options.verboseP) console.log('\ncallback: ' + err0.toString() + '\n' + err0.stack)
-    }
-  })
-
-  if (!options.verboseP) return
-
-  console.log('<<< ' + params.method + ' ' + parts.protocol + '//' + parts.hostname + params.path)
-  underscore.keys(options.headers).forEach((header) => { console.log('<<< ' + header + ': ' + options.headers[header]) })
-  console.log('<<<')
-  if (options.payload) console.log('<<< ' + JSON.stringify(params.payload, null, 2).split('\n').join('\n<<< '))
-}
-
-var runTimeoutId = false
-
-var run = (delayTime) => {
-  if (clientOptions.verboseP) {
-    var entries
-
-    console.log('\nledger client run: clientP=' + (!!client) + ' delayTime=' + delayTime)
-
-    var line = (fields) => {
-      var result = ''
-
-      fields.forEach((field) => {
-        var spaces
-        var max = (result.length > 0) ? 9 : 19
-
-        if (typeof field !== 'string') field = field.toString()
-        if (field.length < max) {
-          spaces = ' '.repeat(max - field.length)
-          field = spaces + field
-        } else {
-          field = field.substr(0, max)
-        }
-        result += ' ' + field
-      })
-
-      console.log(result.substr(1))
-    }
-
-    line([ 'publisher',
-      'blockedP', 'stickyP', 'verified',
-      'excluded', 'eligibleP', 'visibleP',
-      'contribP',
-      'duration', 'visits'
-    ])
-    entries = synopsis.topN() || []
-    entries.forEach((entry) => {
-      var publisher = entry.publisher
-
-      line([ publisher,
-        blockedP(publisher), stickyP(publisher), synopsis.publishers[publisher].options.verified === true,
-        synopsis.publishers[publisher].options.exclude === true, eligibleP(publisher), visibleP(publisher),
-        contributeP(publisher),
-        Math.round(synopsis.publishers[publisher].duration / 1000), synopsis.publishers[publisher].visits ])
-    })
-  }
-
-  if ((typeof delayTime === 'undefined') || (!client)) return
-
-  var active, state, weights, winners
-  var ballots = client.ballots()
-  var data = (synopsis) && (ballots > 0) && synopsisNormalizer()
-
-  if (data) {
-    weights = []
-    data.forEach((datum) => { weights.push({ publisher: datum.site, weight: datum.weight / 100.0 }) })
-    winners = synopsis.winners(ballots, weights)
-  }
-  if (!winners) winners = []
-
-  try {
-    winners.forEach((winner) => {
-      var result
-
-      if (!contributeP(winner)) return
-
-      result = client.vote(winner)
-      if (result) state = result
-    })
-    if (state) muonWriter(pathName(statePath), state)
-  } catch (ex) {
-    console.log('ledger client error(2): ' + ex.toString() + (ex.stack ? ('\n' + ex.stack) : ''))
-  }
-
-  if (delayTime === 0) {
-    try {
-      delayTime = client.timeUntilReconcile()
-    } catch (ex) {
-      delayTime = false
-    }
-    if (delayTime === false) delayTime = random.randomInt({ min: msecs.minute, max: 10 * msecs.minute })
-  }
-  if (delayTime > 0) {
-    if (runTimeoutId) return
-
-    active = client
-    if (delayTime > (1 * msecs.hour)) delayTime = random.randomInt({ min: 3 * msecs.minute, max: msecs.hour })
-
-    runTimeoutId = setTimeout(() => {
-      runTimeoutId = false
-      if (active !== client) return
-
-      if (!client) return console.log('\n\n*** MTR says this can\'t happen(1)... please tell him that he\'s wrong!\n\n')
-
-      if (client.sync(callback) === true) return run(0)
-    }, delayTime)
-    return
-  }
-
-  if (client.isReadyToReconcile()) return client.reconcile(uuid.v4().toLowerCase(), callback)
-
-  console.log('what? wait, how can this happen?')
-}
-
-var getStateInfo = (state) => {
-  var ballots, i, transaction
-  var info = state.paymentInfo
-  var then = underscore.now() - msecs.year
-
-  if (!state.properties.wallet) return
-
-  ledgerInfo.paymentId = state.properties.wallet.paymentId
-  ledgerInfo.passphrase = state.properties.wallet.keychains.passphrase
-
-  ledgerInfo.created = !!state.properties.wallet
-  ledgerInfo.creating = !ledgerInfo.created
-
-  ledgerInfo.reconcileFrequency = state.properties.days
-  ledgerInfo.reconcileStamp = state.reconcileStamp
-
-  if (info) {
-    ledgerInfo._internal.paymentInfo = info
-    cacheReturnValue()
-  }
-
-  ledgerInfo.transactions = []
-  if (!state.transactions) return updateLedgerInfo()
-
-  for (i = state.transactions.length - 1; i >= 0; i--) {
-    transaction = state.transactions[i]
-    if (transaction.stamp < then) break
-
-    if ((!transaction.ballots) || (transaction.ballots.length < transaction.count)) continue
-
-    ballots = underscore.clone(transaction.ballots || {})
-    state.ballots.forEach((ballot) => {
-      if (ballot.viewingId !== transaction.viewingId) return
-
-      if (!ballots[ballot.publisher]) ballots[ballot.publisher] = 0
-      ballots[ballot.publisher]++
-    })
-
-    ledgerInfo.transactions.push(underscore.extend(underscore.pick(transaction,
-                                                                   [ 'viewingId', 'contribution', 'submissionStamp', 'count' ]),
-                                                                   { ballots: ballots }))
-  }
-
-  observeTransactions(state.transactions)
-  updateLedgerInfo()
-}
-
-// Observe ledger client state.transactions for changes.
-// Called by getStateInfo(). Updated state provided by ledger-client.
-var cachedTransactions = null
-var observeTransactions = (transactions) => {
-  if (underscore.isEqual(cachedTransactions, transactions)) {
-    return
-  }
-  // Notify the user of new transactions.
-  if (getSetting(settings.PAYMENTS_NOTIFICATIONS) && cachedTransactions !== null) {
-    const newTransactions = underscore.difference(transactions, cachedTransactions)
-    if (newTransactions.length > 0) {
-      const newestTransaction = newTransactions[newTransactions.length - 1]
-      showNotificationPaymentDone(newestTransaction.contribution.fiat)
-    }
-  }
-  cachedTransactions = underscore.clone(transactions)
-}
-
-var balanceTimeoutId = false
-
-var getBalance = () => {
-  if (!client) return
-
-  balanceTimeoutId = setTimeout(getBalance, 1 * msecs.minute)
-  if (!ledgerInfo.address) return
-
-  if (!ledgerBalance) ledgerBalance = require('ledger-balance')
-  ledgerBalance.getBalance(ledgerInfo.address, underscore.extend({ balancesP: true }, client.options),
-  (err, provider, result) => {
-    var unconfirmed
-    var info = ledgerInfo._internal.paymentInfo
-
-    if (err) return console.warn('ledger balance warning: ' + JSON.stringify(err, null, 2))
-
-    if (typeof result.unconfirmed === 'undefined') return
-
-    if (result.unconfirmed > 0) {
-      unconfirmed = (result.unconfirmed / 1e8).toFixed(4)
-      if ((info || ledgerInfo).unconfirmed === unconfirmed) return
-
-      ledgerInfo.unconfirmed = unconfirmed
-      if (info) info.unconfirmed = ledgerInfo.unconfirmed
-      if (clientOptions.verboseP) console.log('\ngetBalance refreshes ledger info: ' + ledgerInfo.unconfirmed)
-      return updateLedgerInfo()
-    }
-
-    if (ledgerInfo.unconfirmed === '0.0000') return
-
-    if (clientOptions.verboseP) console.log('\ngetBalance refreshes payment info')
-    getPaymentInfo()
-  })
-}
-
-var logError = (err, caller) => {
-  if (err) {
-    ledgerInfo.error = {
-      caller: caller,
-      error: err
-    }
-    console.error('Error in %j: %j', caller, err)
-    return true
-  } else {
-    ledgerInfo.error = null
-    return false
-  }
-}
-
-var getPaymentInfo = () => {
-  var amount, currency
-
-  if (!client) return
-
-  try {
-    ledgerInfo.bravery = client.getBraveryProperties()
-    if (ledgerInfo.bravery.fee) {
-      amount = ledgerInfo.bravery.fee.amount
-      currency = ledgerInfo.bravery.fee.currency
-    }
-
-    client.getWalletProperties(amount, currency, function (err, body) {
-      var info = ledgerInfo._internal.paymentInfo || {}
-
-      if (logError(err, 'getWalletProperties')) {
-        return
-      }
-
-      info = underscore.extend(info, underscore.pick(body, [ 'buyURL', 'buyURLExpires', 'balance', 'unconfirmed', 'satoshis' ]))
-      info.address = client.getWalletAddress()
-      if ((amount) && (currency)) {
-        info = underscore.extend(info, { amount: amount, currency: currency })
-        if ((body.rates) && (body.rates[currency])) {
-          info.btc = (amount / body.rates[currency]).toFixed(8)
-        }
-      }
-      ledgerInfo._internal.paymentInfo = info
-      updateLedgerInfo()
-      cacheReturnValue()
-    })
-  } catch (ex) {
-    console.error('properties error: ' + ex.toString())
-  }
-}
-
-var setPaymentInfo = (amount) => {
-  var bravery
-
-  if (!client) return
-
-  try { bravery = client.getBraveryProperties() } catch (ex) {
-// wallet being created...
-
-    return setTimeout(function () { setPaymentInfo(amount) }, 2 * msecs.second)
-  }
-
-  amount = parseInt(amount, 10)
-  if (isNaN(amount) || (amount <= 0)) return
-
-  underscore.extend(bravery.fee, { amount: amount })
-  client.setBraveryProperties(bravery, (err, result) => {
-    if (err) return console.error('ledger setBraveryProperties: ' + err.toString())
-
-    if (result) muonWriter(pathName(statePath), result)
-  })
-  if (ledgerInfo.created) getPaymentInfo()
-}
-
-var cacheReturnValue = () => {
-  var chunks, cache, paymentURL
-  var info = ledgerInfo._internal.paymentInfo
-
-  if (!info) return
-
-  if (!ledgerInfo._internal.cache) ledgerInfo._internal.cache = {}
-  cache = ledgerInfo._internal.cache
-
-  paymentURL = 'bitcoin:' + info.address + '?amount=' + info.btc + '&label=' + encodeURI('Brave Software')
-  if (cache.paymentURL === paymentURL) return
-
-  cache.paymentURL = paymentURL
-  updateLedgerInfo()
-  try {
-    chunks = []
-
-    qr.image(paymentURL, { type: 'png' }).on('data', (chunk) => { chunks.push(chunk) }).on('end', () => {
-      cache.paymentIMG = 'data:image/png;base64,' + Buffer.concat(chunks).toString('base64')
-      updateLedgerInfo()
-    })
-  } catch (ex) {
-    console.error('qr.imageSync error: ' + ex.toString())
-  }
-}
-
-var networkConnected = underscore.debounce(() => {
-  if (!client) return
-
-  if (runTimeoutId) {
-    clearTimeout(runTimeoutId)
-    runTimeoutId = false
-  }
-  if (client.sync(callback) === true) run(random.randomInt({ min: msecs.minute, max: 10 * msecs.minute }))
-
-  if (balanceTimeoutId) clearTimeout(balanceTimeoutId)
-  balanceTimeoutId = setTimeout(getBalance, 5 * msecs.second)
-}, 1 * msecs.minute, true)
 
 /*
  * low-level utilities
  */
 
-var muonWriter = (path, payload) => {
-  muon.file.writeImportant(path, JSON.stringify(payload, null, 2), (success) => {
-    if (!success) return console.error('write error: ' + path)
 
-    if ((quitP) && (!getSetting(settings.PAYMENTS_ENABLED)) && (getSetting(settings.SHUTDOWN_CLEAR_HISTORY))) {
-      if (ledgerInfo._internal.debugP) console.log('\ndeleting ' + path)
-      return fs.unlink(path, (err) => { if (err) console.error('unlink error: ' + err.toString()) })
-    }
-
-    if (ledgerInfo._internal.debugP) console.log('\nwrote ' + path)
-  })
-}
-
-var pathName = (name) => {
-  var parts = path.parse(name)
-
-  return path.join(app.getPath('userData'), parts.name + parts.ext)
-}
-
-/*
- * UI controller functionality
- */
-
-const showNotifications = () => {
-  if (getSetting(settings.PAYMENTS_ENABLED)) {
-    if (getSetting(settings.PAYMENTS_NOTIFICATIONS)) showEnabledNotifications()
-  } else {
-    showDisabledNotifications()
-  }
-}
-
-// When Payments is disabled
-const showDisabledNotifications = () => {
-  if (!getSetting(settings.PAYMENTS_NOTIFICATION_TRY_PAYMENTS_DISMISSED)) {
-    const firstRunTimestamp = appStore.getState().get('firstRunTimestamp')
-    if (new Date().getTime() - firstRunTimestamp < appConfig.payments.delayNotificationTryPayments) {
-      return
-    }
-    notificationTryPaymentsMessage = locale.translation('notificationTryPayments')
-    appActions.showNotification({
-      greeting: locale.translation('updateHello'),
-      message: notificationTryPaymentsMessage,
-      buttons: [
-        {text: locale.translation('noThanks')},
-        {text: locale.translation('notificationTryPaymentsYes'), className: 'primaryButton'}
-      ],
-      options: {
-        style: 'greetingStyle',
-        persist: false
-      }
-    })
-  }
-}
-
-/**
-* Show message that it's time to add funds if reconciliation is less than
-* a day in the future and balance is too low.
-* 24 hours prior to reconciliation, show message asking user to review
-* their votes.
-*/
-const showEnabledNotifications = () => {
-  const reconcileStamp = ledgerInfo.reconcileStamp
-
-  if (!reconcileStamp) return
-
-  if (reconcileStamp - underscore.now() < msecs.day) {
-    if (sufficientBalanceToReconcile()) {
-      if (shouldShowNotificationReviewPublishers()) {
-        showNotificationReviewPublishers(reconcileStamp + ((ledgerInfo.reconcileFrequency - 2) * msecs.day))
-      }
-    } else if (shouldShowNotificationAddFunds()) {
-      showNotificationAddFunds()
-    }
-  } else if (reconcileStamp - underscore.now() < 2 * msecs.day) {
-    if (sufficientBalanceToReconcile() && (shouldShowNotificationReviewPublishers())) {
-      showNotificationReviewPublishers(underscore.now() + msecs.day)
-    }
-  }
-}
-
-const sufficientBalanceToReconcile = () => {
-  const balance = Number(ledgerInfo.balance || 0)
-  const unconfirmed = Number(ledgerInfo.unconfirmed || 0)
-  return ledgerInfo.btc &&
-    (balance + unconfirmed > 0.9 * Number(ledgerInfo.btc))
-}
-
-const shouldShowNotificationAddFunds = () => {
-  const nextTime = getSetting(settings.PAYMENTS_NOTIFICATION_ADD_FUNDS_TIMESTAMP)
-  return !nextTime || (underscore.now() > nextTime)
-}
-
-const showNotificationAddFunds = () => {
-  const nextTime = underscore.now() + (3 * msecs.day)
-  appActions.changeSetting(settings.PAYMENTS_NOTIFICATION_ADD_FUNDS_TIMESTAMP, nextTime)
-
-  addFundsMessage = addFundsMessage || locale.translation('addFundsNotification')
-  appActions.showNotification({
-    greeting: locale.translation('updateHello'),
-    message: addFundsMessage,
-    buttons: [
-      {text: locale.translation('turnOffNotifications')},
-      {text: locale.translation('updateLater')},
-      {text: locale.translation('addFunds'), className: 'primaryButton'}
-    ],
-    options: {
-      style: 'greetingStyle',
-      persist: false
-    }
-  })
-}
-
-const shouldShowNotificationReviewPublishers = () => {
-  const nextTime = getSetting(settings.PAYMENTS_NOTIFICATION_RECONCILE_SOON_TIMESTAMP)
-  return !nextTime || (underscore.now() > nextTime)
-}
-
-const showNotificationReviewPublishers = (nextTime) => {
-  appActions.changeSetting(settings.PAYMENTS_NOTIFICATION_RECONCILE_SOON_TIMESTAMP, nextTime)
-
-  reconciliationMessage = reconciliationMessage || locale.translation('reconciliationNotification')
-  appActions.showNotification({
-    greeting: locale.translation('updateHello'),
-    message: reconciliationMessage,
-    buttons: [
-      {text: locale.translation('turnOffNotifications')},
-      {text: locale.translation('dismiss')},
-      {text: locale.translation('reviewSites'), className: 'primaryButton'}
-    ],
-    options: {
-      style: 'greetingStyle',
-      persist: false
-    }
-  })
-}
-
-// Called from observeTransactions() when we see a new payment (transaction).
-const showNotificationPaymentDone = (transactionContributionFiat) => {
-  notificationPaymentDoneMessage = locale.translation('notificationPaymentDone')
-    .replace(/{{\s*amount\s*}}/, transactionContributionFiat.amount)
-    .replace(/{{\s*currency\s*}}/, transactionContributionFiat.currency)
-  // Hide the 'waiting for deposit' message box if it exists
-  appActions.hideNotification(addFundsMessage)
-  appActions.showNotification({
-    greeting: locale.translation('updateHello'),
-    message: notificationPaymentDoneMessage,
-    buttons: [
-      {text: locale.translation('turnOffNotifications')},
-      {text: locale.translation('Ok'), className: 'primaryButton'}
-    ],
-    options: {
-      style: 'greetingStyle',
-      persist: false
-    }
-  })
-}
 
 module.exports = {
   init: init,
