+++ js/constants/appConstants.js
@@ -33,9 +33,7 @@ const appConstants = {
   APP_ON_CLEAR_BROWSING_DATA: _,
   APP_IMPORT_BROWSER_DATA: _,
   APP_UPDATE_LEDGER_INFO: _,
-  APP_LEDGER_RECOVERY_STATUS_CHANGED: _,
   APP_UPDATE_LOCATION_INFO: _,
-  APP_UPDATE_PUBLISHER_INFO: _,
   APP_SHOW_NOTIFICATION: _, /** @param {Object} detail */
   APP_HIDE_NOTIFICATION: _, /** @param {string} message */
   APP_BACKUP_KEYS: _,
@@ -111,8 +109,6 @@ const appConstants = {
   APP_URL_BAR_SUGGESTIONS_CHANGED: _,
   APP_SEARCH_SUGGESTION_RESULTS_AVAILABLE: _,
   APP_DEFAULT_SEARCH_ENGINE_LOADED: _,
-  APP_CHANGE_LEDGER_PINNED_PERCENTAGES: _,
-  APP_ENABLE_UNDEFINED_PUBLISHERS: _,
   APP_TAB_PINNED: _,
   APP_DRAG_STARTED: _,
   APP_DRAG_ENDED: _,
